# Term-5/DevelopNetworkApplication/DevelopmentNetworkApplicationBackend/VikingsServer/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Viking's hikes",
        "title": "VIKINGS",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/mightyK1ngRichard",
            "email": "dimapermyakov55@gmai.com"
        },
        "version": "1.0"
    },
    "host": "localhost:7070",
    "basePath": "/",
    "paths": {
...
        "/api/v3/destination-hikes": {
            "put": {
                "description": "Обновление порядкого номера города в истории похода",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пункты походов"
                ],
                "summary": "Обновление порядка локаций похода",
                "parameters": [
                    {
                        "description": "Данные для добавления города в поход",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.UpdateDestinationHikeNumberReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated Destination Hike ID",
                        "schema": {
                            "$ref": "#/definitions/ds.UpdateDestinationHikeNumberRes"
                        }
                    },
                    "400": {
                        "description": "Плохой запрос",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResp"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибку",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResp"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаление локации из истории похода по идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Пункты походов"
                ],
                "summary": "Удаление локации из похода",
                "parameters": [
                    {
                        "description": "Идентификатор локации в походе",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.DeleteDestinationToHikeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Удаленный идентификатор локации",
                        "schema": {
                            "$ref": "#/definitions/ds.DeleteDestinationToHikeRes"
                        }
                    },
                    "400": {
                        "description": "Плохой запрос",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResp"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ds.DeleteDestinationToHikeReq": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "ds.DeleteDestinationToHikeRes": {
            "type": "object",
            "properties": {
                "deleted_destination_hike": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "ds.UpdateDestinationHikeNumberReq": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "serial_number": {
                    "type": "integer"
                }
            }
        },
        "ds.UpdateDestinationHikeNumberRes": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "handler.errorResp": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Описание ошибки"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}